+++ packages/admin/src/pages/Pipeline/index.tsx
@@ -110,7 +110,7 @@ export default function () {
       extra={
         <Button
           onClick={() => {
-            window.open('https://vanblog.mereith.com/features/pipeline.html', '_blank');
+            window.open('https://github.com/CKboss/vanblog/blob/dev/dsh/docs/features/pipeline.md', '_blank');
           }}
         >
           帮助文档